requirements.txt
face_recognition==1.3.0
matplotlib==3.5.3
numpy==1.19.5
Pillow==9.4.0
tensorflow==2.5.0
tensorflow_intel==2.11.0
Flask~=2.2.3
requests~=2.28.2